+++ src/modules/wifi/wifi.c
@@ -132,8 +132,9 @@ void ffGenerateWifiJsonResult(FF_MAYBE_UNUSED FFWifiOptions* options, yyjson_mut
         yyjson_mut_val* conn = yyjson_mut_obj_add_obj(doc, obj, "conn");
         yyjson_mut_obj_add_strbuf(doc, conn, "status", &wifi->conn.status);
         yyjson_mut_obj_add_strbuf(doc, conn, "ssid", &wifi->conn.ssid);
-        yyjson_mut_obj_add_strbuf(doc, conn, "macAddress", &wifi->conn.macAddress);
+        yyjson_mut_obj_add_strbuf(doc, conn, "bssid", &wifi->conn.macAddress);
         yyjson_mut_obj_add_strbuf(doc, conn, "protocol", &wifi->conn.protocol);
+        yyjson_mut_obj_add_strbuf(doc, conn, "security", &wifi->conn.security);
         yyjson_mut_obj_add_real(doc, conn, "signalQuality", wifi->conn.signalQuality);
         yyjson_mut_obj_add_real(doc, conn, "rxRate", wifi->conn.rxRate);
         yyjson_mut_obj_add_real(doc, conn, "txRate", wifi->conn.txRate);
@@ -158,7 +159,7 @@ void ffPrintWifiHelpFormat(void)
         "Interface status",
         "Connection status",
         "Connection SSID",
-        "Connection mac address",
+        "Connection BSSID",
         "Connection protocol",
         "Connection signal quality (percentage)",
         "Connection RX rate",