+++ CHANGELOG.md
@@ -16,10 +16,12 @@ Features:
 * Support st terminal font detection (TerminalFont, Linux)
 * Add option `--migrate-config <?target-file-path>`
 * Support Nvidia GPU temp detection via nvml (GPU)
+* Try supporting Wifi authentication type detection in macOS Sonoma. Please file a feature request if you get `to be supported (num)` with result of `/System/Library/PrivateFrameworks/Apple80211.framework/Resources/airport -I | grep auth` (Wifi, macOS)
 
 Bugfixes:
 * Better GPU memory and type detection (GPU, Windows)
 * Don't print display type twice (Display)
+* Detect BSSID instead of Wifi MAC address to align with other platforms (Wifi, macOS)
 
 Logos:
 * Add Black Mesa